Bug Fix: Issue #12
Description: Calls to CGDisplayPixelsWidth and CGDisplayPixelsHeight stalls when passed an integer value of 0 as the argument. This issue also occurs in when calling the CGDisplayPixelWide with 0 as the argument in Objective C. So the issue is not with pyobjc. So I believe the issue is a result in a change in the implementation of the underlying Objective-C library. Below is sample code in Objective-C which reproduces the issue.

```
#import <Foundation/Foundation.h>

int main(int argc, const char * argv[]) {
    CGDisplayPixelsWide(0);
    return 0;
}
```

Solution: The simplest solution was to substitute the problematic argument (0) with Quarts.CGMainDisplayID(), which returns the display ID for the main monitor. This function is the recommended way of providing the CGDisplay with the display ID.