Refactoring Mouse tests and making sure they work here.
My desktop has two differently-sized monitors, with this layout:

          .--------.
    .-----|        |
    |     |        |
    '-----'--------'

This means the top-left region of the desktop is not accessible. This
means that position 0,0 can't be reached. This means that all tests
failed.

This commit changes the test position to the screen center, which should
be available on all systems. In addition, some systems have *hot
corners*, with actions assigned to the screen corners. It is better to
avoid triggering those actions when running tests.

This commit also changes `assertTrue` to `assertEqual`.