Rewritten/refactored the tweening code.
I started doing this when the test was broken, later I discovered the
bug was in the test (see the previous commit).

Anyway, this commit improves the tweening motion by removing one loop
(replacing by a direct calculation) and by pre-calculating the path and
adding the final coordinates there (instead of the duplicate code handling
it afterwards).

It also enables the division operation to always use float (as in Python
3).