Update _window_win.py
I updated the close method because CloseWindow using the C wrapper actually minimizes the selected window. 
I imported win32con to enable the ability to send messages to other windows. This is because generally windows cannot be closed from programs that are running on a different thread.

The updated close method will attempt to close the window by sending the window a message to perform the close event.
In the case of a program that pops up a Save Changes dialog, the dialog will appear and the window will not automatically close.

I added a closeNow method that will also attempt to close the window by sending the window a message, however, this is the destroy event.
In the case of a program that usually pops up a Save Changes dialog, the dialog will NOT appear and the window will automatically close.